Questionable post implementation
Can now post updates to existing fragments.
This is done by checking if the ID of the paragraph is 'newParagraph'.

This will need to be reviewed. It was initially going to check if id
was defined. However javascript got confused about this.

Included a check so that focus can be taken when needed by paragraphs.
The problem was calling createParagraph from search result rendering.
This stole focus when unhelpful.

Need to have handling of id undefined, may still appear. Though I think
I have removed the cause.